orchestrator/requirements.txt
pyyaml>=6.0
redis>=5.0
docker>=6.0
requests>=2.31
python-dotenv>=1.0
aiohttp>=3.12
